refactor(ui): create shared ApiKeyInput component
